requirements.txt
torch
transformers
beir
colbert-ai
scikit-learn
numpy
pandas
tqdm
matplotlib
seaborn
pyyaml
jupyter
